requirements.txt
pytest==7.4.0
requests==2.31.0
allure-pytest==2.13.2
pytest-html==3.2.0
selenium~=4.16.0
python-dotenv~=1.0.0
pyperclip~=1.8.2